+++ include/geos/operation/overlayng/EdgeSourceInfo.h
@@ -40,7 +40,7 @@ class GEOS_DLL EdgeSourceInfo {
 private:
 
     // Members
-    int index;
+    uint8_t index;
     int dim;
     bool edgeIsHole;
     int depthDelta;
@@ -48,10 +48,10 @@ private:
 
 public:
 
-    EdgeSourceInfo(int p_index, int p_depthDelta, bool p_isHole);
-    EdgeSourceInfo(int p_index);
+    EdgeSourceInfo(uint8_t p_index, int p_depthDelta, bool p_isHole);
+    EdgeSourceInfo(uint8_t p_index);
 
-    int getIndex() const { return index; }
+    uint8_t getIndex() const { return index; }
     int getDimension() const { return dim; }
     int getDepthDelta() const { return depthDelta; }
     bool isHole() const { return edgeIsHole; }
